fabric event params modified
